+++ tests/ui/deprecations.rs
@@ -3,6 +3,7 @@
 use pyo3::prelude::*;
 
 #[pyclass]
+#[text_signature = "()"]
 struct TestClass {
     num: u32,
 }
@@ -14,20 +15,24 @@ impl TestClass {
     const DEPRECATED_NAME_CONSTANT: i32 = 0;
 
     #[name = "num"]
+    #[text_signature = "()"]
     fn deprecated_name_pymethod(&self) { }
 
     #[staticmethod]
     #[name = "custom_static"]
+    #[text_signature = "()"]
     fn deprecated_name_staticmethod() {}
 }
 
 #[pyfunction]
 #[name = "foo"]
+#[text_signature = "()"]
 fn deprecated_name_pyfunction() { }
 
-#[pymodule]
+#[pymodule(deprecated_module_name)]
 fn my_module(_py: Python, m: &PyModule) -> PyResult<()> {
     #[pyfn(m, "some_name")]
+    #[text_signature = "()"]
     fn deprecated_name_pyfn() { }
 
     Ok(())